Managed to create a query tog et all tasks and a mutation to add a task
